components: fix deadlock at components' destructor
At the destructor of V4L2DecodeComponent and
V4L2EncodeComponent, we post a task to the delegated thread to clean
up the thread and wait for the task done. Originally, we assume the
destructors do not run on the delegated thread that is created inside
the component. However, if we initiate a shared_ptr by
weak_from_this() on the delegated thread when the client is about to
release the shared_ptr, the destructor will be called on the delegated
thread. Then deadlock occurs when waiting for the task done.

To fix the deadlock, this CL avoid posting a task to shutdown the
delegated thread if the destructor is called on the delegated thread.

Bug: 204055006
Test: android.media.cts.AdaptivePlaybackTest

Change-Id: I3997e2568a57d6c23161784ad032dadc7ed81d7c
(cherry picked from commit 863f63751b9eaa26fa8fc0e45c9a2e6c7d5204d4)