v4l2_codec2 encoder: Remove 'VIDEO_ENCODER' flag workaround.
In the past we avoided using the 'VIDEO_ENCODER' flag on input buffers.
Due to format conversion these buffers would never be sent directly to
the host, and adding the flag would cause an extra shadow buffer to be
allocated on the host. This is no longer the case and no performance
differences can be observed when always adding the 'VIDEO_ENCODER'
flag, so this CL removes the workaround.

Bug: 167640667
Test: arc.VideoEncodeAccel.h264_192p_i420_vm
Change-Id: Id247b62dfe2e6d09bc5dccbf050b0b4fe8cc263f
(cherry picked from commit a354b9103fe6b3eb303fa9eb33ee8cc0d8f36835)