v4l2_codec2 encoder: Support zero-copy for RGB frames in converter.
The cleanup in ag/16211950 makes it easy to now support zero-copy in
the format converter for RGB video frames. This is currently not used
as the encoder input format is NV12, but we can support if for future
use cases.

Bug: 205203831
Test: arc.VideoEncodeAccel.h264_192p_i420_vm
Change-Id: I4ccc64f7a5b8f401b8df9f2f3205f091e40bbf06
(cherry picked from commit 19161c4c3531fb78cb688d05e57ce892aa85fff0)