V4L2Encoder: fix support for single-buffer formats
The V4L2 encoder driver currently reports the number of NV12 memory
planes incorrectly: whereas NV12 is a single-buffer format (which
multi-planar counterpart is NV12M), the driver reports as many memory
planes as there are component planes.

We are about to fix this with http://crrev.com/c/3358090, but right now
the V4L2Encoder does not manage single-buffer formats properly: the
detection of such formats is based on the number of planes in the frame
(which will always be the number of color planes), and the V4L2
parameters are not properly set, notably the plane size.

Fix this by using the mMultiPlanar member of the input layout to detect
single-buffer formats, and properly set the V4L2 queue inputs in that
case. This allows the V4L2Encoder to operate successfully both when
http://crrev.com/c/3358090 is applied and when it is not.

Bug: b:212214039
Test: arc.VideoEncodeAccel.h264_192p_i420_vm passes on zork-arc-r without http://crrev.com/c/3358090.
Test: arc.VideoEncodeAccel.h264_192p_i420_vm passes on zork-arc-r with http://crrev.com/c/3358090.

Change-Id: Id79828f570851f8bba8ff16a897617a15dac9a19
(cherry picked from commit cbec391821f95c88ee501c85aa38d0672ad3efd8)