Reland "V4L2DevicePoller: Always poll V4L2 events"
This is a reland of commit 8f6a260e89bf1b33b6bcdd2803b653f5361db77a

This version of the CL includes a fix for b/253993076.  The root cause of
the regression was caused by not calling poll again with
POLLIN | POLLOUT | POLLPRI after schedulePoll() was called and
only POLLPRI was dequeued. It was fixed by setting mPollBuffers to true if
it was set to true before the poll and the poll did not yielded that buffers
are pending.

Original change's description:
> This CL changes V4L2DevicePoller behaviour. From now on, V4L2 events are
> polled continuously and schedulePoll will schedule poll for buffer
> related events. V4L2Device::poll method was modified to accept a
> parameter with option to decide if POLLIN | POLLOUT events are enabled
> for poll (POLLERR | POLLPRI are always enabled).
>
> Bug: 230794283
> Test: V4L2_EVENT_SOURCE_CHANGE is received correctly \
>    during android.media.cts.DecoderTest#testEOSBehaviorVP8
> Change-Id: I4f063795bb0d6d27c186aaf8730fef3babd0ef8e

Bug: 230794283
Bug: 253993076
Test: android.media.cts.DecoderTest#testEOSBehaviorVP8
Test: android.video.cts.VideoEncoderDecoderTest
Change-Id: I0578d4ca440736ea858a951e4c048f0ac4e7bbcf
(cherry picked from commit e5de4ffa0524651c97938f310d26ac5b733ebc1f)