V4L2Decoder: do not check for DRC for secure codec
It's not possible to access the frame data if the secure codec is used.

An attempt to map() such data will result in an error, and attempt to
wait for the DRC will cause a nullptr dereference.

This patch adds a workaround, which skips the DRC verification if the
secure codec is used.

Bug: 280853786
Bug: 274864105
Test: GtsExoPlayerTestCases
Change-Id: I5b9d990c87b8afef2817f6bb9b9d8afb42017ca9
(cherry picked from commit 8fce6a632586b49aa23b14235363dca0f1fe0a7c)