V4L2DecodeInterface: Increase output delay for VPx
The decoder might held a few frames as a reference for decoding. Since
Android T the Codec2 is more prone to timeout the component if one is not
producing frames. This might especially occur when those frames are held
for reference and playback/decoding is paused. With increased output delay
we inform Codec2 not to timeout the component, if number of frames in
components is less then the number of maximum reference frames that
could be held by decoder.

This CL increases C2_PARAMKEY_OUTPUT_DELAY to maximum number of frames
that VPx specs specifies.

Bug: 277456388
Test: YouTube VP9 video remains paused after pause
Change-Id: I2ba1aa3ef806c0d48888964577287d4c4afc4052