V4L2DecodeInterface: Align maxInputSize with client request
This CL takes into consideration client request for the
max-input-size. Usually this request was smaller and
component recommended a hardcoded value for
area < 1920*1088 but it proved to be not enough for
some specific case. First the parameter is read and
stored, then we compare this value with the one
calculated as before and we finally set the biggest one.
In this way another hardcoding of arbitrary size
for area between 1920*1080 and 1920*1088 is avoided.

Bug: 280733224
Test: CtsMediaEncoderTestCases
Change-Id: I9486a6d81f9dcdd66ade8bfd10e5b4545ffcaf44