V4L2ComponentName: Fix codec name assertions
The V4L2ComponentName::isValid() is often used as an assertion in other
functions. In some of them, the codec name could be passed as
std::string, and in some others the const char *.
The type discrepancy was causing a build errors when the LOG_NDEBUG was
defined.

The simplest fix is to align the codec name type to be std::string
everywhere.

Test: m
Change-Id: Ib4f6c7f0273ee57d889cfe2e60f0926c349a5871
(cherry picked from commit 42d307dad3e564089a25436aecbf28d8e6206e91)