v4l2: Increase the buffer size of bitstream
The allocated buffer size for the encoded bitstream is always set to 1MB
which might not be always accurate. It is possible to exceed it when
high bitrates are used, like 10Mb/s. The encoder might put zeros
to the output buffer to meet required bitstream. Also in case of slow
devices the encoder might put zeros to meet the time constraints.
For the sake of compatibility, set it to the same values as the Chrome
browser does for a 1080p stream.

Bug: 326164993
Test: CtsMediaCodecTestCases
Change-Id: I4c2d06da588d2a12a593bbba3461384d8b3dffca